% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_khaldoon.R
\name{print.khaldoon_test}
\alias{print.khaldoon_test}
\title{Print Method for Khaldoon Test Results}
\usage{
\method{print}{khaldoon_test}(x, ...)
}
\arguments{
\item{x}{An object of class 'khaldoon_test'}

\item{...}{Additional parameters}
}
\value{
No return value, called for side effects
}
\description{
Print Method for Khaldoon Test Results
}
