% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sample_random_data}
\alias{sample_random_data}
\title{Sample Random Data}
\usage{
sample_random_data(n = 100, mean = 0, sd = 1)
}
\arguments{
\item{n}{Number of observations (default: 100)}

\item{mean}{Mean of the noise (default: 0)}

\item{sd}{Standard deviation of the noise (default: 1)}
}
\value{
Numeric vector of white noise
}
\description{
Generate random white noise data for testing purposes
}
