% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_khaldoon.R
\name{plot.khaldoon_test}
\alias{plot.khaldoon_test}
\title{Plot Method for Khaldoon Test Results}
\usage{
\method{plot}{khaldoon_test}(x, which = "all", ...)
}
\arguments{
\item{x}{An object of class "khaldoon_test"}

\item{which}{Which plots to display. Can be 1, 2, 3, or "all" (default: "all")
1 = Power Spectrum
2 = Bootstrap Distribution
3 = Cumulative Spectrum}

\item{...}{Additional graphical parameters}
}
\description{
Produces diagnostic plots for Khaldoon periodicity test results including
power spectrum, bootstrap distribution, and QQ plot.
}
\examples{
set.seed(123)
x <- sin(2 * pi * seq(0, 10, length.out = 200)) + rnorm(200, sd = 0.3)
result <- khaldoon_periodicity(x)
plot(result)
plot(result, which = 1)  # Only power spectrum

}
