% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/khaldoon_periodicity.R
\name{print.khaldoon_test}
\alias{print.khaldoon_test}
\title{Print Method for Khaldoon Test}
\usage{
\method{print}{khaldoon_test}(x, ...)
}
\arguments{
\item{x}{An object of class "khaldoon_test"}

\item{...}{Additional arguments (not used)}
}
\description{
Print Method for Khaldoon Test
}
