% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_khaldoon.R
\name{summary.khaldoon_test}
\alias{summary.khaldoon_test}
\title{Summary Method for Khaldoon Test Results}
\usage{
\method{summary}{khaldoon_test}(object, ...)
}
\arguments{
\item{object}{An object of class 'khaldoon_test'}

\item{...}{Additional parameters}
}
\value{
No return value, called for side effects
}
\description{
Summary Method for Khaldoon Test Results
}
